// src copy/types/twitchAPI.ts
type SubscriptionStatus =
  | "enabled"
  | "webhook_callback_verification_pending"
  | "webhook_callback_verification_failed"
  | "notification_failures_exceeded"
  | "authorization_revoked"
  | "moderator_removed"
  | "user_removed"
  | "version_removed"
  | "beta_maintenance"
  | "websocket_disconnected"
  | "websocket_failed_ping_pong"
  | "websocket_received_inbound_traffic"
  | "websocket_connection_unused"
  | "websocket_internal_error"
  | "websocket_network_timeout"
  | "websocket_network_error";

type ShardStatus =
  | "enabled"
  | "webhook_callback_verification_pending"
  | "webhook_callback_verification_failed"
  | "notification_failures_exceeded"
  | "websocket_disconnected"
  | "websocket_failed_ping_pong"
  | "websocket_received_inbound_traffic"
  | "websocket_internal_error"
  | "websocket_network_timeout"
  | "websocket_network_error";

type TransportMethod = "webhook" | "websocket" | "conduit";

interface Transport {
  method: TransportMethod;
  callback?: string;
  session_id?: string;
  conduit_id?: string;
  connected_at?: string;
  disconnected_at?: string;
}

export interface GetEventSubSubscriptionsRequest {
  status?: SubscriptionStatus;
  type: EventSubTopics;
  user_id?: string;
  after?: string;
}

export interface GetEventSubSubscriptionsResponse {
  data: {
    id: string;
    status: SubscriptionStatus;
    type: string;
    version: string;
    condition: object; // This could be further detailed if the structure of condition objects is known.
    created_at: string;
    transport: Transport;
    cost: number;
  }[]
}
// ...
interface User {
  id: string;
  login: string;
  display_name: string;
  type: "admin" | "global_mod" | "staff" | "";
  broadcaster_type: "affiliate" | "partner" | "";
  description: string;
  profile_image_url: string;
  offline_image_url: string;
  email?: string;
  created_at: string; // Should be converted to a proper date type if needed
}

export interface getUserResponse {
  data: User[];
}